+++ api/dfg/CornellDiningNow.py
@@ -3,7 +3,7 @@ import requests
 from api.dfg.DfgNode import DfgNode
 
 from api.datatype.Eatery import Eatery
-from api.util import dining_id_to_internal_id, CORNELL_DINING_URL
+from util.constants import dining_id_to_internal_id, CORNELL_DINING_URL
 
 from datetime import date
