+++ api/dfg/EateriesFromDB.py
@@ -5,15 +5,15 @@ import re
 import pytz
 
 from api.datatype.Eatery import Eatery, EateryID
-from api.datatype.EateryException import EateryException
+from api.datatype.EateryAlert import EateryAlert
 from api.datatype.Event import Event
 from api.datatype.Menu import Menu
 from api.datatype.MenuCategory import MenuCategory
 from api.datatype.MenuItem import MenuItem
 from api.dfg.DfgNode import DfgNode
 
-from eateries.models import EateryStore, ExceptionStore
-from eateries.serializers import EateryStoreSerializer, ExceptionStoreSerializer
+from eateries.models import EateryStore, AlertStore
+from eateries.serializers import EateryStoreSerializer, AlertStoreSerializer
 
 from datetime import datetime
 
@@ -25,17 +25,17 @@ class EateriesFromDB(DfgNode):
         eateries = EateryStore.objects.all()
         serialized_eateries = EateryStoreSerializer(data=eateries, many=True)
         serialized_eateries.is_valid()
-        exceptions = ExceptionStore.objects.filter(end_timestamp__gte=datetime.now().timestamp(), start_timestamp__lte=datetime.now().timestamp())
-        serialized_exceptions = ExceptionStoreSerializer(data=exceptions, many=True)
-        serialized_exceptions.is_valid()
-        return list(map(lambda x: EateriesFromDB.eatery_from_serialized(x, serialized_exceptions.data), serialized_eateries.data))
+        alerts = AlertStore.objects.filter(end_timestamp__gte=datetime.now().timestamp(), start_timestamp__lte=datetime.now().timestamp())
+        serialized_alerts = AlertStoreSerializer(data=alerts, many=True)
+        serialized_alerts.is_valid()
+        return list(map(lambda x: EateriesFromDB.eatery_from_serialized(x, serialized_alerts.data), serialized_eateries.data))
 
     @staticmethod
     def none_repr(str):
         return None if str == None or len(str) == 0 else str
 
     @staticmethod
-    def eatery_from_serialized(serialized_eatery: dict, serialized_exceptions: list[dict]) -> Eatery:
+    def eatery_from_serialized(serialized_eatery: dict, serialized_alerts: list[dict]) -> Eatery:
         return Eatery(
             id=EateryID(serialized_eatery["id"]),
             name=EateriesFromDB.none_repr(serialized_eatery["name"]),
@@ -49,7 +49,7 @@ class EateriesFromDB(DfgNode):
             location=EateriesFromDB.none_repr(serialized_eatery["location"]),
             online_order=serialized_eatery["online_order"],
             online_order_url=EateriesFromDB.none_repr(serialized_eatery["online_order_url"]),
-            exceptions = EateriesFromDB.exceptions(serialized_eatery["id"], serialized_exceptions)
+            alerts = EateriesFromDB.alerts(serialized_eatery["id"], serialized_alerts)
         )
     
     @staticmethod
@@ -64,16 +64,16 @@ class EateriesFromDB(DfgNode):
         return pay_methods
 
     @staticmethod
-    def exceptions(eatery_id: int, serialized_exceptions: list[dict]):
-        return [EateriesFromDB.exception_from_serialized(exception) for exception in serialized_exceptions if exception["eatery"] == eatery_id]
+    def alerts(eatery_id: int, serialized_alerts: list[dict]):
+        return [EateriesFromDB.alert_from_serialized(alert) for alert in serialized_alerts if alert["eatery"] == eatery_id]
 
     @staticmethod
-    def exception_from_serialized(serialized_exception: dict):
-        return EateryException(
-            id=serialized_exception["id"],
-            description=serialized_exception["description"],
-            start_timestamp=serialized_exception["start_timestamp"],
-            end_timestamp=serialized_exception["end_timestamp"]
+    def alert_from_serialized(serialized_alert: dict):
+        return EateryAlert(
+            id=serialized_alert["id"],
+            description=serialized_alert["description"],
+            start_timestamp=serialized_alert["start_timestamp"],
+            end_timestamp=serialized_alert["end_timestamp"]
         )
