+++ api/dfg/system/EateryGenerator.py
