+++ api/views.py
@@ -9,11 +9,11 @@ from api.dfg.EateryStubs import EateryStubs
 from api.dfg.EateriesFromDB import EateriesFromDB
 from api.dfg.macros.EateryEvents import EateryEvents
 
-from api.dfg.util.DictResponseWrapper import DictResponseWrapper
-from api.dfg.util.ConvertToJson import ConvertToJson
-from api.dfg.util.EateryGenerator import EateryGenerator
-from api.dfg.util.InMemoryCache import InMemoryCache
-from api.dfg.util.Mapping import Mapping
+from api.dfg.system.DictResponseWrapper import DictResponseWrapper
+from api.dfg.system.ConvertToJson import ConvertToJson
+from api.dfg.system.EateryGenerator import EateryGenerator
+from api.dfg.system.InMemoryCache import InMemoryCache
+from api.dfg.system.Mapping import Mapping
 
 from api.dfg.macros.LeftMergeEateries import LeftMergeEateries
