+++ eateries/admin.py
@@ -7,7 +7,7 @@ admin.site.register(models.EateryStore)
 admin.site.register(models.MenuStore)
 admin.site.register(models.ItemStore)
 admin.site.register(models.SubItemStore)
-admin.site.register(models.ExceptionStore)
+admin.site.register(models.AlertStore)
 admin.site.register(models.CategoryStore)
 admin.site.register(models.CategoryItemAssociation)
 admin.site.register(models.DayOfWeekEventSchedule)
