+++ eateries/management/commands/ingest_db_snapshot.py
@@ -1,10 +1,8 @@
 
 from django.core.management.base import BaseCommand
-
-from eateries.util import SnapshotFileName
+from util.constants import SnapshotFileName
 
 import eateries.serializers as serializers
-
 import json
 
 class Command(BaseCommand):
@@ -18,25 +16,16 @@ class Command(BaseCommand):
             for line in file:
                 if (len(line) > 2):
                     json_objs.append(json.loads(line))
-            serialized_objs = serializer(data=json_objs, many=True)
+            serialized_objs = serializer(data=json~_objs, many=True)
             serialized_objs.is_valid()
             serialized_objs.save()
 
     def handle(self, *args, **options):
         self.ingest_data(serializers.EateryStoreSerializer, SnapshotFileName.EATERY_STORE)
-        self.ingest_data(serializers.ExceptionStoreSerializer, SnapshotFileName.EXCEPTION_STORE)
+        self.ingest_data(serializers.AlertStoreSerializer, SnapshotFileName.ALERT_STORE)
         self.ingest_data(serializers.MenuStoreSerializer, SnapshotFileName.MENU_STORE)
         self.ingest_data(serializers.CategoryStoreSerializer, SnapshotFileName.CATEGORY_STORE)
         self.ingest_data(serializers.ItemStoreSerializer, SnapshotFileName.ITEM_STORE)
         self.ingest_data(serializers.SubItemStoreSerializer, SnapshotFileName.SUBITEM_STORE)
         self.ingest_data(serializers.CategoryItemAssociationSerializer, SnapshotFileName.CATEGORY_ITEM_ASSOCIATION)
         self.ingest_data(serializers.DayOfWeekEventScheduleSerializer, SnapshotFileName.DAY_OF_WEEK_EVENT_SCHEDULE)
-
-        # with open(f"{folder_path}/{SnapshotFileName.EXCEPTION_STORE.value}", "r") as file:
-        #     serialized_objs = []
-        #     for line in file:
-        #         if (len(line) > 2):
-        #             serialized_objs.append(json.loads(line))
-        #     exception_objs = serializers.ExceptionStoreSerializer(data=serialized_objs, many=True)
-        #     exception_objs.is_valid()
-        #     eatery_objs.save()
